secret-init

+++ internal/scope/cluster.go
@@ -2,6 +2,6 @@ package scope
 
 type Cluster struct {
 	Name                     string `json:"name"`
-	ControlplaneMachineCount int    `json:"controlplaneMachineCount"`
-	WorkerMachineCount       int    `json:"workerMachineCount"`
+	ControlplaneMachineCount int64  `json:"controlplaneMachineCount"`
+	WorkerMachineCount       int64  `json:"workerMachineCount"`
 }
